Make the em dash (—) also work when using options
Some devices (such as iPhones) automatically change it to `—` when you type `--`. To get it to stay as `--` you have to type something like `-f-` and then delete the f, which can be annoying. Some people probably don't realize this problem and wonder why it didn't work. Allowing both should make for a better user experience.