refactor: Use native WS client with image servers
This allows dropping a dependency from the bot itself, but the `ws` module is still required on the image server because Node.js still does not have a native WS server.

Also bumps the required Node.js version to 22.4.0, as this version is the first to include a native WS client.