feat: Add role-based tag gating
Tags are now only allowed to be created by members with the Administrator permission by default. The role subcommands allow for adding, removing, and listing extra roles to allow more members to create tags.

This also includes a few other fixes related to tags and database handling, as I found/fixed them on the same branch as this was being worked on and they were a bit too intertwined with the changes here to separate.