fix: Use raw fetch for direct image sending
Oceanic's client handles some network errors by sending them as events; however, it does not also throw them in the function context, and it is nearly impossible to tie them to specific function calls. In worst-case scenarios, this can result in jobs stalling for excessive amounts of time - which in turn can result in issues with application commands.

By using direct `fetch` calls, we can bypass this issue entirely. There are some things we lose out on here, such as latency compensation and ratelimit handling, but these requests are proxied in production anyways and smaller environments are unlikely to hit those ratelimits so there's not too much of a downside.