feat: Directory-based subcommands
To create a subcommand, create a regular command file to act as the base, then create a new directory with the same name as your command next to it and place your subcommand files in there.

Single-depth subcommands have been tested and are working, however subcommand groups are still untested (and probably do not work at the moment).