ci: Automatically mirror to Codeberg
A new Codeberg mirror has been created for the esmBot repo: https://codeberg.org/esm/esmBot

Really not happy with GitHub's leadership lately. I'm sick of them embedding garbage "AI" features in nearly every area of the platform, as well as the (now former as of today) CEO trying to say that devs that aren't playing into the hype cycle should get out of the tech industry. It's unlikely that this project will move 100% off of GitHub due to various factors (even if the Codeberg repo becomes the main one pull requests will still probably be accepted from GitHub for accessibility reasons), but it doesn't hurt to have a backup plan.

Mirroring is configured on both sides, so commits to the Codeberg repo should also be mirrored to GitHub and vice versa. As such, pull requests are accepted on both forges.